backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
pytesseract
python-docx
python-pptx
transformers
torch
sentencepiece
